pyproject.toml
[project]
name = "jwks-server"
version = "0.1.0"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "pyjwt[crypto]>=2.9",
  "cryptography>=42.0",
  "httpx>=0.27",
  "pytest>=8.3",
  "pytest-cov>=5.0"
]

[tool.pytest.ini_options]
addopts = "-q --cov=app --cov-report=term-missing"
